MessageEnquerer implemented
