+++ source/Infrastructure/OutgoingMessages/MessageEnqueuer.cs
@@ -1,0 +1,40 @@
+﻿// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System;
+using Application.OutgoingMessages;
+using Domain.OutgoingMessages;
+using Domain.OutgoingMessages.Queueing;
+
+namespace Infrastructure.OutgoingMessages;
+
+public class MessageEnqueuer
+{
+    private readonly IOutgoingMessageStore _outgoingMessageStore;
+
+    public MessageEnqueuer(IOutgoingMessageStore outgoingMessageStore)
+    {
+        _outgoingMessageStore = outgoingMessageStore;
+    }
+
+    public void Enqueue(OutgoingMessage message)
+    {
+        ArgumentNullException.ThrowIfNull(message);
+
+        var actorMessageQueue = ActorMessageQueue.CreateFor(message.Receiver, BusinessReason.From(message.BusinessReason));
+        actorMessageQueue.Enqueue(message);
+
+        _outgoingMessageStore.Add(message);
+    }
+}
